feat: Implement dynamic column widths and minimum quantity threshold
- Updated ViewportManager to support dynamic column widths for quantity and price.
- Enhanced MarketDataSimulator to include a minimum quantity threshold for order generation.
- Modified CanvasRenderer to calculate and render dynamic column widths based on maximum quantities and prices.
- Adjusted PriceLadder to filter out zero and near-zero quantities based on the new threshold.
- Introduced default minimum quantity threshold in types and configuration.
- Updated Webpack configuration for improved development experience with source maps.

+++ src/core/Interop/WasmExports.cs
@@ -86,7 +86,7 @@ public partial class WasmExports
     /// Process a single price level update
     /// </summary>
     [JSExport]
-    public static void ProcessPriceLevelUpdate(int side, double price, int quantity, int numOrders)
+    public static void ProcessPriceLevelUpdate(int side, double price, double quantity, int numOrders)
     {
         ExecuteSafely(nameof(ProcessPriceLevelUpdate), () =>
         {
@@ -95,7 +95,7 @@ public partial class WasmExports
             var update = new PriceLevel(
                 (Side)side,
                 (decimal)price,
-                quantity,
+                (decimal)quantity,
                 numOrders
             );
 
@@ -107,7 +107,7 @@ public partial class WasmExports
     /// Process a single price level update without auto-flushing
     /// </summary>
     [JSExport]
-    public static void ProcessPriceLevelUpdateNoFlush(int side, double price, int quantity, int numOrders)
+    public static void ProcessPriceLevelUpdateNoFlush(int side, double price, double quantity, int numOrders)
     {
         ExecuteSafely(nameof(ProcessPriceLevelUpdateNoFlush), () =>
         {
@@ -116,7 +116,7 @@ public partial class WasmExports
             var update = new PriceLevel(
                 (Side)side,
                 (decimal)price,
-                quantity,
+                (decimal)quantity,
                 numOrders
             );
 
@@ -157,7 +157,7 @@ public partial class WasmExports
         int orderId,
         int side,
         double price,
-        int quantity,
+        double quantity,
         int priority,
         int updateType,
         int isOwnOrder)
@@ -170,7 +170,7 @@ public partial class WasmExports
                 orderId,
                 (Side)side,
                 (decimal)price,
-                quantity,
+                (decimal)quantity,
                 priority,
                 isOwnOrder != 0
             );
@@ -187,7 +187,7 @@ public partial class WasmExports
         int orderId,
         int side,
         double price,
-        int quantity,
+        double quantity,
         int priority,
         int updateType,
         int isOwnOrder)
@@ -200,7 +200,7 @@ public partial class WasmExports
                 orderId,
                 (Side)side,
                 (decimal)price,
-                quantity,
+                (decimal)quantity,
                 priority,
                 isOwnOrder != 0
             );
